ecommerce/src/main/java/com/example/practise/spring/service/implementations/CustomerServiceImpl.java
package com.example.practise.spring.service.implementations;

import com.example.practise.spring.entity.Customer;
import com.example.practise.spring.repository.CustomerRepository;
import com.example.practise.spring.service.interfaces.CustomerService;
import org.springframework.stereotype.Service;
import java.util.List;

@Service
public class CustomerServiceImpl implements CustomerService {

    private final CustomerRepository customerRepository;

    public CustomerServiceImpl(CustomerRepository customerRepository) {
        this.customerRepository = customerRepository;
    }

    public List<Customer> getCustomers() {
        return customerRepository.findAll();
    }

    public Customer addCustomer(Customer customer) {
        return customerRepository.save(customer);
    }


}
